# openapi/openapi.yaml
openapi: 3.0.0
...
info:
  description: PCIT API
  version: 1.0.0
  title: PCIT API
  termsOfService: "https://github.com/pcit-ce/pcit/blob/master/docs/terms-of-service.md"
  contact:
    email: ci@khs1994.com
    name: pcit
    url: "https://ci.khs1994.com"
  license:
    name: MIT
    url: "https://github.com/pcit-ce/pcit/blob/master/LICENSE"
...
paths:
...
  "/repo/{username}/{repo.name}/requests":
    get:
      tags:
        - Repositories.requests
      summary: get repo request list
      description: This will return a list of requests belonging to a repository.
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - name: repo.name
          in: path
          required: true
          schema:
            type: string
      operationId: get-repo-request-list
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo-builds"
              examples:
                example1:
                  value:
                    - id: "112"
                      branch: gh-pages
                      commit_id: 0fa67f2b1c0d59998b71035b402849f5075a3c60
                      tag: null
                      commit_message: "Deploy Git by PCIT https://ci.khs1994.com"
                      compare: >-
                        https://github.com/pcit-ce/pcit/compare/d492689de936...0fa67f2b1c0d
                      committer_name: pcit-ce
                      committer_username: pcit-deploy-robot
                      build_status: misconfigured
                      event_type: push
                      pull_request_number: null
                      created_at: "1577892876"
                      finished_at: null
                    - id: "111"
                      branch: master
                      commit_id: f807efaa677757682aa7a3f226d3e799086e90dd
                      tag: null
                      commit_message: |-
                        Support actions

                        Signed-off-by: Kang HuaiShuai
                      compare: >-
                        https://github.com/pcit-ce/pcit/compare/6746c883155e...f807efaa6777
                      committer_name: Kang HuaiShuai
                      committer_username: khs1994
                      build_status: queued
                      event_type: push
                      pull_request_number: null
                      created_at: "1577892253"
                      finished_at: null
    post:
      tags:
        - Repositories.requests
      summary: create repo single request
      description: >-
        This will create a request for an individual repository, triggering a
        build to run on PCIT.
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - name: repo.name
          in: path
          required: true
          schema:
            type: string
      operationId: create-repo-single-request
      responses:
        "200":
          description: success
  "/repo/{username}/{repo.name}/request/{request.id}":
    get:
      tags:
        - Repositories.requests
      summary: get repo single request
      description: This will return information about an individual request.
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - name: repo.name
          in: path
          required: true
          schema:
            type: string
        - name: request.id
          in: path
          required: true
          schema:
            type: integer
      operationId: get-repo-single-request
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/request"
              examples:
                example1:
                  value:
                    id: "111"
                    git_type: github
                    rid: "119219872"
                    event_type: push
                    build_status: queued
                    branch: master
                    tag: null
                    pull_request_title: null
                    pull_request_number: null
                    pull_request_source: null
                    compare: >-
                      https://github.com/pcit-ce/pcit/compare/6746c883155e...f807efaa6777
                    commit_id: f807efaa677757682aa7a3f226d3e799086e90dd
                    commit_message: |-
                      Support actions

                      Signed-off-by: Kang HuaiShuai
                    committer_uid: null
                    committer_name: Kang HuaiShuai
                    committer_username: khs1994
                    committer_email: khs1994@khs1994.com
                    committer_pic: null
                    author_uid: null
                    author_name: Kang HuaiShuai
                    author_username: khs1994
                    author_email: khs1994@khs1994.com
                    author_pic: null
                    created_at: "1577892253"
                    finished_at: null
                    deleted_at: null
                    config: "{}"
                    action: null
                    check_suites_id: null
                    internal: "1"
                    private: "0"
                    unique_key: "0"
...
  /validate:
    post:
      tags:
        - System
      summary: validate .pcit.yml
      description: validate .pcit.yml
      operationId: validate
      requestBody:
        content:
          plain/text:
            schema:
              type: string
            examples:
              example1:
                value: |-
                  language: php

                  steps:
                    script: echo exec script
          application/json:
            schema:
              type: string
            examples:
              example:
                value: |-
                  {"language": "php","steps":{"script":"echo exec script"}}
      responses:
        "200":
          description: validate result
          content:
            plain/text:
              schema:
                type: string
                description: validate passed
              examples:
                example1:
                  value: "ok"
            application/json:
              schema:
                type: array
                description: validate error
                items:
                  type: object
              examples:
                example1:
                  value:
                    - steps: The property steps is required
                    - "": The property lanuage is not defined and the definition does not allow additional properties
                    - "": The property test is not defined and the definition does not allow additional properties
components:
...
  schemas:
...
    repo-builds:
      title: repo build list
      description: repo build list
      type: object
      properties:
        id:
          type: integer
          description: builds id
        branch:
          type: string
          description: trigger branch
        commit_id:
          type: string
          description: commit sha
        tag:
          type: string
          description: tag name
        commit_message:
          type: string
          description: commit message
        compare:
          type: string
          description: compare url
        commiter_name:
          type: string
          description: committer name
        committer_username:
          type: string
          description: committer username
        build_status:
          type: string
          description: build status
        event_type:
          type: string
          description: event type
        pull_request_number:
          type: integer
          description: pull request number
        created_at:
          type: integer
          description: created time
        finished_at:
          type: integer
          description: finished time
...
    request:
      title: request
      type: object
      properties:
        id:
          type: integer
          description: build id
        git_type:
          type: string
          description: git type
        rid:
          type: integer
          description: repo id
        event_type:
          type: string
          description: event type
        build_status:
          type: string
          description: build status
        branch:
          type: string
          description: branch
        tag:
          type: string
          description: tag name
        pull_request_title:
          type: string
          description: pull request title
        pull_request_number:
          type: integer
          description: pull request number
        pull_request_source:
          type: string
          description: pull request source repo
        compare:
          type: string
          description: compare url
        commit_id:
          type: string
          description: commit sha
        commit_message:
          type: string
          description: commit message
        committer_uid:
          type: integer
          description: committer uid
        committer_name:
          type: string
          description: committer name
        committer_username:
          type: string
          description: committer username
        committer_email:
          type: string
          description: committer email
        committer_pic:
          type: string
          description: committer pic
        author_uid:
          type: integer
          description: author uid
        author_name:
          type: string
          description: author name
        author_username:
          type: string
          description: author username
        author_email:
          type: string
          description: author email
        author_pic:
          type: string
          description: author pic
        created_at:
          type: integer
          description: create time
        finished_at:
          type: integer
          description: finish time
        deleted_at:
          type: integer
          description: delete time
        config:
          type: string
          description: pcit config string
        action:
          type: string
          description: action
        check_suites_id:
          type: integer
          description: check suites id
        internal:
          type: integer
          description: internal
        private:
          type: boolean
          description: is private
        unique_key:
          type: string
          description: unique key
...
  securitySchemes:
    BearerToken:
      type: apiKey
      description: Bearer Token
      in: header
      name: Authorization
...
security:
  - BearerToken: []
...